fix(stampedtemplate): sublcass DocumentFragment
This follows the spec[1] section 3.1 which states that
`TemplateInstance` is inherited from `DocumentFragment`.

This makes the StampedTemplate a lot more ergonomic as it no longer
needs to hold a reference to a document fragment, because it is one. It
also makes it more ergonomic to use as you simply
`appendChild(stampedtemplate)` rather than getting the fragment part.

1: https://github.com/w3c/webcomponents/blob/gh-pages/proposals/Template-Instantiation.md#31-basics