Extend `processPropertyIdentity` to handle Element
Closes [github/template-parts#62][]

Extends the `processPropertyIdentity` function from the `processors`
module to pass-through `Element` instances, instead of coercing them
through calls to `String()`. In order to accommodate that change, this
commit also extends the `TemplatePart` interface to accept `Element`
instances for its `value:` property.

The change to the `createTreeWalker` invocation within `collectParts`
was highlighted by TypeScript, as the final `boolean` argument doesn't
conform to the [createTreeWalker][] signature. It was a necessary change
to build the project and execute the test suite locally.

[github/template-parts#62]: https://github.com/github/template-parts/issues/62
[createTreeWalker]: https://developer.mozilla.org/en-US/docs/Web/API/Document/createTreeWalker